+++ src/components/Stamper/style.css
