+++ src/components/machine.js
@@ -3,6 +3,7 @@ import { Container, Row, Col } from "react-bootstrap";
 
 import * as constants from "../constants.js";
 
+import Stamper from "./Stamper";
 import BiscuitLifecycle from "./biscuit/biscuitLifecycle.js";
 import BoostSwitch from "./boostSwitch.js";
 import BiscuitContainer from "./container.js";
@@ -10,7 +11,6 @@ import Conveyor from "./conveyor.js";
 import Extruder from "./extruder.js";
 import Motor from "./motor.js";
 import Oven from "./oven.js";
-import Stamper from "./stamper/stamper.js";
 import Switch from "./switch.js";
 
 const Machine = () => {
